+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 (CI) [#7](https://github.com/0xPellNetwork/pelldvs/pull/7) ci: add changelog check in CI  
 
+(chore) [#11](https://github.com/0xPellNetwork/pelldvs/pull/11) chore: change emulator default port for e2e test
+
 ### Bug Fixes
 
 ## v0.2.2